RAG/requirements.txt
flask==2.2.3
pandas==2.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
langchain==0.0.267
sqlalchemy==2.0.7
flask-cors==3.0.10
PyPDF2==3.0.1
tabula-py==2.7.0
requests==2.28.2
openpyxl==3.1.1
jpype1==1.4.1
reportlab==3.6.12